requirements.txt
torch==2.2.0
scikit-learn==1.3.1
xgboost==2.0.0
catboost==1.2.2
torch==2.2.0
deepchem==2.7.1
# torch-cluster==1.6.3
torch-geometric==2.4.0
# torch-scatter==2.1.2
# torch-sparse==0.6.18
# torch-spline-conv==1.2.2
rdkit==2023.3.3
stmol==0.0.9
streamlit==1.33.0
streamlit-option-menu==0.3.12
pandas==1.3.0
mlflow==2.13.0
mordredcommunity==2.0.5
py3dmol==2.1.0
optuna==3.3.0
plotly==5.20.0
joblib==1.4.0
joblib-progress==1.0.5
scipy==1.8.0
scopy==1.2.5
seaborn==0.13.2
matplotlib==3.8.4
ipython_genutils==0.2.0